+++ test/constants/copyright.spec.ts
@@ -1,0 +1,8 @@
+import { expect } from "chai";
+import copyrightedDictionaries from "../../src/constants/copyright.js";
+
+describe("The copyright object", () => {
+  it("is unmodifiable.", () => {
+    expect(copyrightedDictionaries).to.be.frozen;
+  });
+});